Split prosopography ops into domain modules
